+++ src/SDK/Client/Platform/methods/contracts/get.spec.ts
@@ -1,4 +1,5 @@
 import {expect} from 'chai';
+import getResponseMetadataFixture from '../../../../../test/fixtures/getResponseMetadataFixture';
 import get from "./get";
 import identitiesFixtures from "../../../../../../tests/fixtures/identities.json";
 import contractsFixtures from "../../../../../../tests/fixtures/contracts.json";
@@ -6,32 +7,44 @@ import DataContractFactory from "@dashevo/dpp/lib/dataContract/DataContractFacto
 import ValidationResult from "@dashevo/dpp/lib/validation/ValidationResult";
 import Identifier from "@dashevo/dpp/lib/Identifier";
 import 'mocha';
-import {ClientApps} from "../../../ClientApps";
+import { ClientApps } from "../../../ClientApps";
+const GetDataContractResponse = require("@dashevo/dapi-client/lib/methods/platform/getDataContract/GetDataContractResponse");
+const NotFoundError = require('@dashevo/dapi-client/lib/transport/GrpcTransport/errors/NotFoundError');
 
 const factory = new DataContractFactory(
+    undefined,
     () => {
         return new ValidationResult();
-    });
+    },
+    () => [42, contractsFixtures.ratePlatform]);
 const dpp = {
-    dataContract: factory
+    dataContract: factory,
+    getProtocolVersion: () => 42,
 }
+factory.dpp = dpp;
 
-const apps = new ClientApps();
+const apps = new ClientApps({
+    ratePlatform: {
+        contractId: contractsFixtures.ratePlatform.$id
+    },
+});
 let client;
 let askedFromDapi;
+let initialize;
 
 describe('Client - Platform - Contracts - .get()', () => {
-    before(()=>{
+    before(function before() {
         askedFromDapi = 0;
         const getDataContract = async (id) => {
             const fixtureIdentifier = Identifier.from(contractsFixtures.ratePlatform.$id);
-            askedFromDapi+=1;
+            askedFromDapi += 1;
 
             if (id.equals(fixtureIdentifier)) {
                 const contract = await dpp.dataContract.createFromObject(contractsFixtures.ratePlatform);
-                return contract.toBuffer()
+                return new GetDataContractResponse(contract.toBuffer(), getResponseMetadataFixture());
             }
-            return null;
+
+            throw new NotFoundError();
         };
 
         client = {
@@ -46,28 +59,36 @@ describe('Client - Platform - Contracts - .get()', () => {
                 return apps
             }
         };
-    })
-    describe('get a contract from string', ()=>{
+
+        initialize = this.sinon.stub();
+    });
+
+    describe('get a contract from string', () => {
         it('should get from DAPIClient if there is none locally', async function () {
 
             // @ts-ignore
-            const contract = await get.call({apps, dpp, client}, contractsFixtures.ratePlatform.$id);
+            const contract = await get.call({apps, dpp, client, initialize}, contractsFixtures.ratePlatform.$id);
             expect(contract.toJSON()).to.deep.equal(contractsFixtures.ratePlatform);
+            expect(contract.getMetadata().getBlockHeight()).to.equal(10);
+            expect(contract.getMetadata().getCoreChainLockedHeight()).to.equal(42);
             expect(askedFromDapi).to.equal(1);
         });
+
         it('should get from local when already fetched once', async function () {
             // @ts-ignore
-            const contract = await get.call({apps, dpp, client}, contractsFixtures.ratePlatform.$id);
+            const contract = await get.call({apps, dpp, client, initialize}, contractsFixtures.ratePlatform.$id);
             expect(contract.toJSON()).to.deep.equal(contractsFixtures.ratePlatform);
+            expect(contract.getMetadata().getBlockHeight()).to.equal(10);
+            expect(contract.getMetadata().getCoreChainLockedHeight()).to.equal(42);
             expect(askedFromDapi).to.equal(1);
         });
     })
 
-    describe('other conditions', ()=>{
+    describe('other conditions', () => {
         it('should deal when contract do not exist', async function () {
             // @ts-ignore
-            const contract = await get.call({apps, dpp, client}, identitiesFixtures.bob.id);
+            const contract = await get.call({apps, dpp, client, initialize}, identitiesFixtures.bob.id);
             expect(contract).to.equal(null);
         });
-    })
+    });
 });