+++ src/SDK/Client/Platform/methods/identities/register.ts
@@ -1,6 +1,8 @@
 import { Platform } from "../../Platform";
-import { wait } from "../../../../../utils/wait";
 import createAssetLockTransaction from "../../createAssetLockTransaction";
+import createIdentityCreateTransition from "./internal/createIdentityCreateTransition";
+import createAssetLockProof from "./internal/createAssetLockProof";
+import broadcastStateTransition from "../../broadcastStateTransition";
 
 /**
  * Register identities to the platform
@@ -8,55 +10,37 @@ import createAssetLockTransaction from "../../createAssetLockTransaction";
  * @param {number} [fundingAmount=10000] - funding amount in duffs
  * @returns {Identity} identity - a register and funded identity
  */
-export default async function register(this: Platform, fundingAmount : number = 10000): Promise<any> {
-    const { client, dpp } = this;
+export default async function register(
+  this: Platform,
+  fundingAmount : number = 10000
+): Promise<any> {
+    await this.initialize();
+
+    const { client } = this;
 
     const account = await client.getWalletAccount();
 
     const {
         transaction: assetLockTransaction,
-        privateKey: assetLockPrivateKey
+        privateKey: assetLockPrivateKey,
+        outputIndex: assetLockOutputIndex
     } = await createAssetLockTransaction(this, fundingAmount);
 
     // Broadcast Asset Lock transaction
     await account.broadcastTransaction(assetLockTransaction);
+    const assetLockProof = await createAssetLockProof(this, assetLockTransaction, assetLockOutputIndex);
 
-    // Wait some time for propagation
-    await wait(1000);
+    const { identity, identityCreateTransition, identityIndex } = await createIdentityCreateTransition(
+        this, assetLockProof, assetLockPrivateKey
+    );
+    await broadcastStateTransition(this, identityCreateTransition);
 
-    // Create Identity
-    const assetLockOutPoint = assetLockTransaction.getOutPointBuffer(0);
-
-    const identityIndex = await account.getUnusedIdentityIndex();
-
-    // @ts-ignore
-    const { privateKey: identityPrivateKey } = account.getIdentityHDKeyByIndex(identityIndex, 0);
-    const identityPublicKey = identityPrivateKey.toPublicKey();
-
-    const identity = dpp.identity.create(assetLockOutPoint, [identityPublicKey]);
-
-    // Create ST
-    const identityCreateTransition = dpp.identity.createIdentityCreateTransition(identity);
-
-    identityCreateTransition.signByPrivateKey(assetLockPrivateKey);
-
-    const result = await dpp.stateTransition.validateStructure(identityCreateTransition);
-
-    if (!result.isValid()) {
-        throw new Error(`StateTransition is invalid - ${JSON.stringify(result.getErrors())}`);
-    }
-
-    // Broadcast ST
-    await client.getDAPIClient().platform.broadcastStateTransition(identityCreateTransition.toBuffer());
-
+    // If state transition was broadcast without any errors, import identity to the account
     account.storage.insertIdentityIdAtIndex(
         account.walletId,
         identity.getId().toString(),
         identityIndex,
     );
-
-    // Wait some time for propagation
-    await wait(1000);
 
     return identity;
 }