+++ tests/functional/sdk.js
@@ -36,7 +36,7 @@ describe('SDK', function suite() {
   it('should init a Client', async () => {
     expect(clientInstance.network).to.equal(process.env.NETWORK);
 
-    expect(clientInstance.walletAccountIndex).to.equal(0);
+    expect(clientInstance.defaultAccountIndex).to.equal(0);
 
     expect(clientInstance.getApps().has('dpns')).to.be.true;
     expect(clientInstance.getApps().get('dpns')).to.deep.equal({
@@ -62,7 +62,7 @@ describe('SDK', function suite() {
   })
 
   it('should sign and verify a message', async function () {
-    const idKey = account.getIdentityHDKeyByIndex(0, 0);
+    const idKey = account.identities.getIdentityHDKeyByIndex(0, 0);
     // This transforms from a Wallet-Lib.PrivateKey to a Dashcore-lib.PrivateKey.
     // It will quickly be annoying to perform this, and we therefore need to find a better solution for that.
     const privateKey = Dash.Core.PrivateKey(idKey.privateKey);