+++ CHANGELOG.md
@@ -1,3 +1,161 @@
+## [3.20.3](https://github.com/dashevo/DashJS/compare/v3.20.2...v3.20.3) (2021-08-08)
+
+### Chores
+
+* update DashPay and DPNS contracts ([#240](https://github.com/dashevo/DashJS/issues/240))
+
+
+
+## [3.20.2](https://github.com/dashevo/DashJS/compare/v3.20.1...v3.20.2) (2021-07-28)
+
+
+### Bug Fixes
+
+* InvalidResponse error when connecting to older versions of dapi ([#241](https://github.com/dashevo/DashJS/issues/241))
+
+
+
+## [3.20.1](https://github.com/dashevo/DashJS/compare/v3.20.0...v3.20.1) (2021-07-22)
+
+
+### Features
+
+* update DashPay and DPNS contracts ([4a0f0f8](https://github.com/dashevo/DashJS/commit/4a0f0f84a37fd6f41aa75c718ddd8a5ebc43c452))
+
+
+
+# [3.20.0](https://github.com/dashevo/DashJS/compare/v3.19.4...v3.20.0) (2021-07-13)
+
+
+### Features
+
+* provide metadata for documents, identities and data contracts ([#231](https://github.com/dashevo/DashJS/issues/231))
+* update wallet options ([#229](https://github.com/dashevo/DashJS/issues/229))
+
+
+### BREAKING CHANGES
+
+* `wallet.defaultAccountIndex` should be used instead of `walletAccountIndex` option
+
+
+
+## [3.19.4](https://github.com/dashevo/DashJS/compare/v3.19.3...v3.19.4) (2021-05-18)
+
+
+### Bug Fixes
+
+* Binary properties in platform queries not encoded properly ([#223](https://github.com/dashevo/js-dash-sdk/pull/223))
+
+## [3.19.3](https://github.com/dashevo/DashJS/compare/v3.19.2...v3.19.3) (2021-05-18)
+
+
+### Bug Fixes
+
+* CBOR not decoding buffers properly in browsers([#219](https://github.com/dashevo/js-dash-sdk/pull/219))
+
+
+## [3.19.2](https://github.com/dashevo/DashJS/compare/v3.19.1...v3.19.2) (2021-05-18)
+
+
+### Bug Fixes
+
+* add some handler for metadata error ([#216](https://github.com/dashevo/DashJS/issues/216))
+
+
+
+## [3.19.1](https://github.com/dashevo/DashJS/compare/v3.19.0...v3.19.1) (2021-05-10)
+
+
+### Bug Fixes
+
+* invalid testnet DPNS contract id ([#214](https://github.com/dashevo/DashJS/issues/214))
+
+
+
+# [3.19.0](https://github.com/dashevo/DashJS/compare/v3.18.2...v3.19.0) (2021-05-04)
+
+### Features
+
+* add `verifyInstantLock` to state repository ([#193](https://github.com/dashevo/DashJS/issues/193))
+* Chain Asset Lock proof ([#203](https://github.com/dashevo/DashJS/issues/203))
+* update to new getStatus endpoint ([#205](https://github.com/dashevo/DashJS/issues/205))
+* remove fake Instant Asset Locks proofs ([#198](https://github.com/dashevo/DashJS/issues/198))
+
+
+### BREAKING CHANGES
+
+* Fallbacks for Instant Asset lock proofs are removed. SDK is not compatible with local network created with mn-bootstrap v0.18
+* See [DPP breaking changes](https://github.com/dashevo/js-dpp/releases/tag/v0.19.0)
+* See [Wallet lib breaking changes](https://github.com/dashevo/wallet-lib/releases/tag/v7.19.0)
+
+
+
+## [3.18.2](https://github.com/dashevo/DashJS/compare/v3.18.1...v3.18.2) (2021-04-28)
+
+### Bug Fixes
+
+* `transaction.isCoinbase` is not a function and other fixes from wallet-lib 7.18.1 ([#236](https://github.com/dashevo/wallet-lib/issues/236))
+
+
+
+## [3.18.1](https://github.com/dashevo/DashJS/compare/v3.18.0...v3.18.1) (2021-03-03)
+
+### Features
+
+* handle wallet async errors ([#190](https://github.com/dashevo/DashJS/issues/190))
+
+
+
+# [3.18.0](https://github.com/dashevo/DashJS/compare/v3.18.0...v3.17.0) (2021-03-03)
+
+### Features
+
+* make identity derivation DIP-11 compatible ([#188](https://github.com/dashevo/DashJS/issues/188)
+* more reliable and secure ST ack ([#180](https://github.com/dashevo/DashJS/issues/180))
+
+
+### Bug Fixes
+
+* ensure name search uses lowercase label ([#183](https://github.com/dashevo/DashJS/issues/183))
+
+
+### BREAKING CHANGES
+
+* Identities registered in versions prior to that one won't sync, as identities are now using hardened derivation and derived from the wallet, not from the account. This change is made to make JS SDK compatible with Android and iOS apps and DIP 11.
+
+
+
+# [3.17.0](https://github.com/dashevo/DashJS/compare/v3.16.2...v3.17.0) (2020-12-30)
+
+
+### Features
+
+* connect to testnet by default ([#176](https://github.com/dashevo/DashJS/issues/176))
+* update `dapi-client`, `dashcore-lib`, `dpp` and `wallet-lib` ([#152](https://github.com/dashevo/DashJS/issues/152), [#172](https://github.com/dashevo/DashJS/issues/172))
+* wait broadcasted data to be available from queries ([#165](https://github.com/dashevo/DashJS/issues/165))
+* asset lock proofs for identity funding ([#169](https://github.com/dashevo/js-dash-sdk/issues/169))
+
+
+### Bug Fixes
+
+* crypto.randomBytes stub causing tests to fail ([#173](https://github.com/dashevo/DashJS/issues/173))
+
+
+### BREAKING CHANGES
+
+* requires a version of DAPI with instant locks implemented, i.e. 0.17 or higher
+
+
+
+## [3.16.2](https://github.com/dashevo/DashJS/compare/v3.16.1...v3.16.2) (2020-11-17)
+
+
+### Bug Fixes
+
+* cannot read property 'getBinaryProperties' of undefined ([#158](https://github.com/dashevo/DashJS/issues/158))
+
+
+
 ## [3.16.1](https://github.com/dashevo/DashJS/compare/v3.16.0...v3.16.1) (2020-10-30)
 
 
@@ -108,7 +266,7 @@
 - **Bug fixes:**
     * fix: wrong assetlock tx fee estimation (#85)
     * fix: generate one-time private key for the asset lock transaction (#86)
-        
+
 ## [3.13.2](https://github.com/dashevo/DashJS/compare/v3.13.1...v3.13.2) (2020-06-12)
 
 - **Bug fixes:**
@@ -146,7 +304,7 @@
   - added replacement of a document. (#41)
   - added deletion of a document (#41)
 
-- **impr**: 
+- **impr**:
   - update to dpp 0.12 (#41)
 
 - **fix**:
@@ -154,7 +312,7 @@
 
 - **Chore, Docs & Tests:**
   - bumped wallet-lib to 6.1 (#41)
- 
+
 # [2.0.0](https://github.com/dashevo/DashJS/compare/v1.1.2...v2.0.0) (2020-03-27)
 
 - **breaking:**
@@ -165,7 +323,7 @@
   - exported file for web environment is now `Dash` instead of `DashJS`
 - **feat**:
   - Sign and verify message (#24)
-- **impr**: 
+- **impr**:
   - Typings documentation (#30)
   - Code cleanup (#31)
   - Export all Dashcore Primitives (under `SDK.Core.*`)
@@ -175,7 +333,7 @@
   - Identity/register: updated getUTXOS usages on (#afda5bbafb940b2e15d5e773d0e8fc5fbc48ee13)
   - fix(StateTransitionBuilder): records type detection (#b49f74b4b8e03e9d1020dd789c62f4310a4fc1ad)
   - broadcasting of contracts (#f4b63e6be692841f1b138e5b058e531a0873f456, #4aa31fec0e5579d7ef8b9222576863a069b95fd3)
-- **chore**: 
+- **chore**:
   - updated for new evonet (updated wallet-lib to 6.0.0)
   -  updated dapi-client to 0.11 (#fba4d55d3281bec5e65605787dd23a6ca3517476)
   - updated DPNS contractID on evonet (#d0cf11d30cf7c9aaef1ffa4a2b8a955fbf5b1184)